Fix signing scheme Detekt violations
