+++ core/apps/src/main/kotlin/sk/styk/martin/apkanalyzer/core/apps/analysis/ApkSigningBlockParser.kt
@@ -1,0 +1,7 @@
+package sk.styk.martin.apkanalyzer.core.apps.analysis
+
+import sk.styk.martin.apkanalyzer.core.apps.model.SigningSchemeVersion
+
+internal interface ApkSigningBlockParser {
+    fun parseSchemeVersions(apkPath: String): Set<SigningSchemeVersion>?
+}
