+++ feature/app-detail/impl/src/main/kotlin/sk/styk/martin/apkanalyzer/feature/appdetail/impl/certificates/CertificatesScreen.kt
@@ -1,3 +1,5 @@
+@file:Suppress("TooManyFunctions")
+
 package sk.styk.martin.apkanalyzer.feature.appdetail.impl.certificates
 
 import android.widget.Toast
